fix some tlvs

+++ konata.core/Msf/Packets/Tlv/T177.cs
@@ -4,10 +4,10 @@ namespace Konata.Msf.Packets.Tlvs
 {
     public class T177 : TlvBase
     {
-        private readonly long _buildTime;
+        private readonly uint _buildTime;
         private readonly string _sdkVersion;
 
-        public T177(long buildTime, string sdkVersion)
+        public T177(uint buildTime, string sdkVersion)
         {
             _buildTime = buildTime;
             _sdkVersion = sdkVersion;
@@ -22,7 +22,7 @@ namespace Konata.Msf.Packets.Tlvs
         {
             StreamBuilder builder = new StreamBuilder();
             builder.PutByte(1);
-            builder.PutUintBE((int)_buildTime);
+            builder.PutUintBE(_buildTime);
             builder.PutString(_sdkVersion);
             return builder.GetBytes();
         }
